refactor(cli): extract paused-event handler from turn-handler.ts
The filesize ratchet flagged GREW: `src/cli/commands/interactive/turn-handler.ts`
grew from its 407-line baseline to 411 lines. This extraction brings it to 365
code lines, comfortably below the baseline, by pulling the usage-limit pause
handler into its own sibling file.

## What moved

`turn-handler.ts` → `turn-handler.paused.ts`:
- `handlePausedEvent()` — named async helper taking explicit parameters
- `PausedPickerRef` type (the abort-controller ref object)
- `PausedEvent` type alias (`Extract<OutputEvent, { type: 'paused' }>`)
- All option-label constants for the interactive picker
- Full JSDoc and `// Contract:`, `// Invariant:` comments

`turn-handler.ts` retains:
- Public call site: `await handlePausedEvent({ event, session, ... })`
- Unchanged public surface — no importer rewrites required
- `pickerRef` typed as `PausedPickerRef` (imported from the new file)
- Removed unused imports: `usageLimitBox`, `runPicker` (both moved)

## Gates

- `pnpm lint` ✓ (TypeScript noEmit, zero errors)
- `pnpm audit:filesize:update` ✓ (baseline updated to 365)
- `pnpm audit:filesize:check` ✓ (all 41 grandfathered files, no new violations)
- `pnpm audit:funcsize:update` ✓
- `pnpm test src/cli/commands/interactive/` ✓ (1145 tests passed)

Before: 411 code lines
After:  365 code lines (−46 from turn-handler.ts, +90 in new file)
