+++ src/cli/commands/interactive/turn-handler.ts
@@ -8,8 +8,6 @@ import { describeForHistory } from '../../input/attachments.js';
 import { recordTurn } from '../../slash/session-stats.js';
 import { palette } from '../../palette.js';
 import { isDebugEnabled } from '../../../utils/debug.js';
-import { usageLimitBox } from '../../render.js';
-import { runPicker } from '../../render/picker.js';
 import { classifyError, presentError } from '../../errors/index.js';
 import {
   type CompletionWriter,
@@ -36,6 +34,7 @@ import { expandAtFileTokens } from './at-file-inject.js';
 import { promoteWithQueuedFlush, previewOneLine } from './queued-flush.js';
 import { createTurnTtfbState, emitPlainTtfbWaiting, observeFirstContent, ttfbRendererOptions } from './turn-handler.ttfb.js';
 import { tickContextProgress } from './turn-handler.context-progress.js';
+import { handlePausedEvent, type PausedPickerRef } from './turn-handler.paused.js';
 
 export { formatToolLine, formatToolResultLine, ToolLane } from './tool-lane.js';
 
@@ -114,7 +113,7 @@ export async function runTurn(
   // narrowing does not collapse the type to `never` in the finally block after
   // the async .then() assignment — a synchronous read in finally always sees
   // the latest write even though the assignment happens in a microtask.
-  const pickerRef: { abort: AbortController | null } = { abort: null };
+  const pickerRef: PausedPickerRef = { abort: null };
   let lastContextProgressMs = 0;
   const toolEvents: ToolEvent[] = [];
   const pendingTools = new Map<string, ToolEvent>();
@@ -482,108 +481,16 @@ export async function runTurn(
         }
 
         if (event.type === 'paused') {
-          // Mark the compositor paused so a submitted line ends the wait (via
-          // the pause-interrupt handler + input-dispatch Enter path) instead of
-          // sitting queued behind the auto-resume. Cleared on resumed / finally.
-          h.setPausedState?.(true);
-          // Disarm before raw console output so the card doesn't tear the
-          // live overlay. Auto-resume path continues — the provider is now
-          // waiting; the stream will deliver a 'resumed' event when ready,
-          // at which point we rebuild a fresh renderer for the replayed turn.
-          await disposeRendererOnce();
-
-          // Invariant: the interactive picker REPLACES the passive card when a
-          // TTY compositor is armed (borrowedCompositor != null) AND the
-          // provider will auto-resume (autoResume === true) — i.e. there is a
-          // live wait to make a decision about. The two are mutually exclusive:
-          // showing both would duplicate the same options (prose card + menu).
-          // Non-TTY surfaces and autoResume=false fall through to the passive
-          // card, which is the only possible surface there.
-          if (borrowedCompositor && event.autoResume === true) {
-            const ac = new AbortController();
-            pickerRef.abort = ac;
-
-            const resetsAtStr = event.resetsAt
-              ? event.resetsAt.toLocaleTimeString(undefined, { hour: '2-digit', minute: '2-digit' })
-              : null;
-            // Option labels are referenced both here and in the .then() match
-            // below — keep them as consts so the definition and the match can
-            // never drift apart.
-            const keepLabel = resetsAtStr
-              ? `Keep waiting — auto-resumes at ${resetsAtStr}`
-              : 'Keep waiting — auto-resume in progress';
-            const switchModelLabel = 'Switch model / provider  (type /model after)';
-            const stopLabel = 'Stop waiting';
-
-            // Contract: the picker header carries the context the passive card
-            // would have shown (limit + reset time) plus the out-of-band
-            // account-switch tip — "switch account" is NOT a selectable option
-            // because it happens via `claude login` in another terminal during
-            // the wait, which the keychain hot-swap picks up automatically.
-            // Options are ordered by increasing disruption so the safe default
-            // (keep waiting) is first and pre-selected.
-            const header = [
-              palette.warning('  ⏳ Usage limit reached.') +
-                (resetsAtStr ? palette.dim(`  Auto-resumes at ${resetsAtStr}.`) : ''),
-              palette.dim('  Tip: run `claude login` in another terminal to switch account — this turn resumes on it automatically.'),
-              '',
-            ];
-
-            void runPicker(borrowedCompositor, {
-              header,
-              options: [keepLabel, switchModelLabel, stopLabel],
-              signal: ac.signal,
-              initialIndex: 0,
-            }).then((result) => {
-              // Picker resolved — null means aborted (resumed/turn-end tore it
-              // down); any result means the user made an explicit choice.
-              pickerRef.abort = null;
-              if (!result) return; // aborted — no action needed
-
-              const choice = result[0];
-              if (choice === undefined || choice === keepLabel) {
-                // Keep waiting (or a defensive undefined): the auto-resume path
-                // continues unchanged; B's Enter-during-pause path stays live.
-                return;
-              }
-
-              // Switch model / Stop waiting: end the wait via the pause-interrupt
-              // path (same mechanism as B's Enter-during-pause). session.interrupt()
-              // ends the stream; pauseInterruptRequested breaks the for-await loop
-              // so recordTurn is skipped and the queued buffer flushes next turn.
-              pauseInterruptRequested = true;
-              if (choice === switchModelLabel) {
-                // Cannot pre-fill the input buffer (no public buffer-set API on
-                // the compositor), so guide with a printed hint instead.
-                (completionWriter ?? { fn: console.log }).fn(
-                  palette.dim('  Hint: type /model <name> to switch, then send your message again.'),
-                );
-              }
-              session.interrupt().catch((err) => {
-                if (isDebugEnabled()) {
-                  console.error('  ' + palette.error('picker pause-interrupt session.interrupt() failed:'), err);
-                }
-              });
-            }).catch((err) => {
-              // Defensive: runPicker never rejects, but the .then() body calls
-              // completionWriter.fn (→ compositor.commitAbove), which can throw
-              // mid-teardown. Swallow outside debug so a throw here can't surface
-              // as an unhandled rejection (the REPL has no process-level handler).
-              if (isDebugEnabled()) {
-                console.error('  ' + palette.error('picker promise rejected:'), err);
-              }
-            });
-          } else {
-            // Passive card — the only surface when no interactive picker applies
-            // (non-TTY, or autoResume=false where there is no wait to decide on).
-            (completionWriter ?? { fn: console.log }).fn(usageLimitBox({
-              reason: event.reason,
-              ...(event.resetsAt !== undefined ? { resetsAt: event.resetsAt } : {}),
-              ...(event.accountId !== undefined ? { accountId: event.accountId } : {}),
-              ...(event.autoResume !== undefined ? { autoResume: event.autoResume } : {}),
-            }));
-          }
-
+          await handlePausedEvent({
+            event,
+            session,
+            borrowedCompositor,
+            pickerRef,
+            completionWriter,
+            disposeRendererOnce,
+            setPausedState: h.setPausedState,
+            onPauseInterrupt: () => { pauseInterruptRequested = true; },
+          });
           continue;
         }
